moved to graph renderer module
