+++ src/app/app.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import * as controlsActions from './actions/controls.actions';
-import * as graphActions from './actions/graph.actions';
 import { ScreenSizeService } from './services/screen-size.service';
 
 @Component({
@@ -16,6 +15,5 @@ export class AppComponent implements OnInit {
 
     ngOnInit() {
         this.store$.dispatch(controlsActions.loadSavedState());
-        this.store$.dispatch(graphActions.initializeGraphSyncProcess());
     }
 }
