+++ src/app/components/error/error.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { requestGraph } from 'src/app/actions/graph.actions';
-import { GraphState } from 'src/app/reducers/graph.reducer';
+import { requestGraph } from 'src/app/modules/graph-renderer/actions';
+import { GraphState } from 'src/app/modules/graph-renderer/reducer';
 
 @Component({
     selector: 'app-error',
