+++ src/app/modules/graph-renderer/actions/graph.actions.ts
@@ -3,9 +3,9 @@ import { createAction, props } from '@ngrx/store';
 import { LndChannel } from 'api/src/models';
 import { ChunkInfo } from 'api/src/models/chunkInfo.interface';
 import { LndNodeWithPosition } from 'src/app/types/node-position.interface';
-import { Chunk } from '../types/chunk.interface';
-import { LnGraph } from '../types/graph.interface';
-import { LndNode } from '../types/node.interface';
+import { Chunk } from '../../../types/chunk.interface';
+import { LnGraph } from '../../../types/graph.interface';
+import { LndNode } from '../../../types/node.interface';
 
 export const requestGraph = createAction('[graph] requestGraph');
