+++ src/app/modules/graph-renderer/components/graph-edge-mesh/index.ts
@@ -1,0 +1,1 @@
+export * from './graph-edge-mesh.component';
