+++ src/app/modules/graph-renderer/components/graph-font-mesh/index.ts
@@ -1,0 +1,1 @@
+export * from './graph-font-mesh.component';
