+++ src/app/modules/graph-renderer/components/graph-node-mesh/index.ts
@@ -1,0 +1,1 @@
+export * from './graph-node-mesh.component';
