+++ src/app/modules/graph-renderer/components/graph-scene/graph-scene.component.scss
