+++ src/app/modules/graph-renderer/components/graph-scene/index.ts
@@ -1,0 +1,1 @@
+export * from './graph-scene.component';
