+++ src/app/modules/graph-renderer/components/index.ts
@@ -1,0 +1,4 @@
+export * from './graph-edge-mesh';
+export * from './graph-font-mesh';
+export * from './graph-node-mesh';
+export * from './graph-scene';
