+++ src/app/modules/graph-renderer/directives/lnd-raycaster-enable.directive.ts
@@ -1,6 +1,7 @@
 import { AfterViewInit, Directive } from '@angular/core';
-import { LndRaycasterService } from '../services/lnd-raycaster-service';
+import { LndRaycasterService } from '../services/lnd-raycaster-service/lnd-raycaster-service';
 
+// eslint-disable-next-line @angular-eslint/directive-selector
 @Directive({ selector: '[lnd-raycaster-enable]' })
 export class LndRaycasterEnableDirective implements AfterViewInit {
     constructor(private raycasterService: LndRaycasterService) {}
