+++ src/app/modules/graph-renderer/directives/lnd-raycaster-scene.directive.ts
@@ -1,6 +1,7 @@
 import { AfterViewInit, Directive, ElementRef } from '@angular/core';
-import { LndRaycasterService } from '../services/lnd-raycaster-service';
+import { LndRaycasterService } from '../services/lnd-raycaster-service/lnd-raycaster-service';
 
+// eslint-disable-next-line @angular-eslint/directive-selector
 @Directive({ selector: '[lnd-raycaster-scene]' })
 export class LndRaycasterSceneDirective implements AfterViewInit {
     constructor(private elRef: ElementRef, private raycasterService: LndRaycasterService) {}
