+++ src/app/modules/graph-renderer/index.ts
@@ -1,0 +1,1 @@
+export * from './graph-renderer.module';
