+++ src/app/modules/graph-renderer/reducer/index.ts
@@ -1,0 +1,1 @@
+export * from './graph.reducer';
