+++ src/app/modules/graph-renderer/shaders/point.ts
