--- src/app/services/.gitignore
@@ -1,1 +1,0 @@
-macaroon.ts
